- added a script to populate the users at server start - a series of debugging while creating the postman collection

+++ src/main/java/com/example/ing/controller/AppointmentController.java
@@ -32,7 +32,7 @@ public class AppointmentController {
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<String> deleteAppointments(@RequestBody List<Integer> appointmentsIds) {
+    public ResponseEntity<String> deleteAppointments(@RequestBody List<Long> appointmentsIds) {
         return appointmentService.deleteAppointments(appointmentsIds);
     }
 }
